Added a retry mechanism for REST API 5xx errors, with linear backoff
remotehttp funnels all REST API calls through a retryable mechanism. It will retry on any internal API errors as well as on 5xx API errors.
If retries are activated, then failed API calls are retried after a pause. The length of the pause increases linearly with the number of attempts for a given request.
Pause base length is configurable per-store, just like number of retries.
This implementation will always read all of the response bodies, even though the bodies are not always used. This is necessary to make the TCP connections reusable under all circumstances. It would be possible to discard rather than read the bodies in some cases.

There is a large number of tests for remotehttp. One test case is currently failing: a PUT operation where the server side first responds with 503, and then succeeds. Under these circumstances, the client will provide an incorrect request body.