remotehttp API allows re-reading the request body multiple times
When HTTP requests fail with a 5xx error, the request - including the body - need to be re-sent.
The caller must therefore not provide an io.Reader for the request body, but an idempotent function which can provide as many io.Readers for the same request body as necessary.

When it comes to remotehttpindex.StoreIndex(), this has been written with the assumption that errors are rare.
When an error occurs, the entire index will be re-written & re-piped for each attempt.
This avoids excessive temporary buffering in the non-error case, but repeats index generation work in the error case.