Add information about the seed indexes. (#199)
This adds the ability to add one or more seed indexes to the `info`
command.

The resulting output will contain the number of chunks, from the index
provided in input, that are available in the seed. Similarly to what
it's already done with the store.

Additionally it also shows the size of the unique (non duplicated)
chunks that are not available from the seed.
This information can be used as a way to roughly estimate the download
size of a particular update. For example if we use:
```
desync info --seed local.caibx -s https://[...]/store update.caibx
```
We will get in output the size of the chunks that are required by
`update.caibx` but that are not already available in our local seed. So
these will be the chunks that we must download from the remote store.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>