Feed multiple chunks to the workers while verifying an index
If we send a single chunk to every Goroutine, the workers will complete
their task immediately and then will wait for a new chunk to validate.

If instead we group together multiple chunks we can improve the overall
performance.

For example, in my test machine, the required time to verify an index
that pointed to a 5gb image took 30 seconds, and with this patch it
dropped to 24 seconds.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>