Add option to discard invalid seeds
If we have a seed index for a file/device that is corrupted we may want
to try harder and continue anyway by discarding the invalid seed and
fallback to the potentially other seeds and/or by just using the store.

This has been added with a new option "--skip-invalid-seeds".

In order to not reduce the installation speed, the process of choosing
and validating the chunks has been refactored, introducing a new concept
called "plan".
We create a plan about where to look for the chunks ahead of time,
before starting to actually write in the destination.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>