Bail out if multiple config entries match a single location
If we have more than one entry in the configuration file that match the
location string that we are currently using (e.g. one with a trailing
slash and another without), we will end up choosing one configuration at
random.

When this situation arise, we should instead abort the execution to
avoid unexpected results.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>