Add new ENV variable to make it easier to parse the progresses
Introduce a new environment variable `DESYNC_ENABLE_PARSABLE_PROGRESS`
that allows us to print in output the progress without the progress bar.

This will make it easier to parse the current completion percentage, and
estimated remaining time, without flooding the journal, or where the
stderr was redirected, with very frequent updates that contain long
progress bars.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>