Validate the written chunks when using a seed
The seed values are validated at the beginning when we create the
extraction "plan".

However a seed might point to a location where is possible to read and
write data. In that case, even if at the beginning we validated the seed
data, the assumption that when we use `WriteInto()` the seed data is
still valid might not be true anymore.

For this reason we should also re-validate the written data to ensure
that what ended up in the destination is exactly what we were expecting.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>