Do not immediately abort the extraction if a seed chunk is invalid
If the seed points to a RW location it may happen that some files change
while we are in the middle of an extraction.

When using the "InvalidSeedActionRegenerate" option, we can try harder
and attempt to take the invalid chunks from the self seed or the store.
If both of those fail too, then we abort the entire operation.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>