Add the ability to specify the seed data path
When using the `--seed foo.caibx` option, Desync expects to find the
data in foo, next to the provided .caibx file.

While this is a valid assumption for files, it requires additional steps
when the data is a block device. For example a common suggestion in
those cases is to create a symlink next to the seed index.

Instead, now `--seed` gains the ability to explicitly set a path for the
data blob. The syntax becomes:
`--seed /path/to/index.caibx:/path/to/data`.

If the colon is missing, the old behavior of expecting the data next to
the index file will still be used by default.

Fixes: #228

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>