Add --error-retry-base-interval as a CLI option
Setting the base interval between error retries is a useful option,
especially when the Internet connection might be unstable.

Given that `error-retry` can already be set via the CLI, by adding an
option for `error-retry-base-interval` too we can avoid the need to
force users to ship a config file when its only purpose was to only
change this value.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>