Add information about the cache
When using `desync info` it might be useful to have additional
information regarding your local cache.

For example knowing the amount of chunks that can be picked up by your
cache can be used as a metric to ensure the cache is actually working as
expected.

With this commit we add `in-cache` to print the number of chunks that
the cache already has, `not-in-seed-nor-cache` with the number of chunks
that needs to be downloaded from the store and finally
`dedup-size-not-in-seed-nor-cache` with the size of the chunks that
needs to be downloaded from the store.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>