Add new `inspect-chunks` command for more detailed chunks info
Add a new `inspect-chunks` command that can provide additional
information about the chunks in an index, including their compressed
size (if a local store is provided).

One use case is for the clients to download the server's inspect-chunks
JSON and feed it into `desync info`, to calculate the precise download
size of an update.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>