Do not call t.Fatal() from non-main goroutines (#291)
Calling t.Fatal() only stops the current goroutine, but not the rest.
The result is different depending on a test, but in the worst case
scenario (like an error in the server part of a test) the client may
hang waiting indefinitely for the response, and the test ends with
a panic from the test timeout, without showing the real reason for
the issue.